refactor : 프로젝트 시작 세팅
